requirements.txt
streamlit
python-docx
langchain
langchain-core
langchain-community
langchain-openai
openai
